+++ apps/backend/src/modules/users/user.repository.ts
@@ -3,19 +3,25 @@ import { type Repository } from "~/libs/types/types.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 import { type UserModel } from "~/modules/users/user.model.js";
 
+import { type UserTaskDay } from "./libs/types/types.js";
 import { type UserDetailsModel } from "./user-details.model.js";
+import { type UserTaskDaysModel } from "./user-task-days.model.js";
 
 class UserRepository implements Repository {
 	private userDetailsModel: typeof UserDetailsModel;
 
 	private userModel: typeof UserModel;
 
+	private userTaskDaysModel: typeof UserTaskDaysModel;
+
 	public constructor(
 		userModel: typeof UserModel,
 		userDetailsModel: typeof UserDetailsModel,
+		userTaskDaysModel: typeof UserTaskDaysModel,
 	) {
 		this.userModel = userModel;
 		this.userDetailsModel = userDetailsModel;
+		this.userTaskDaysModel = userTaskDaysModel;
 	}
 
 	public async create(entity: UserEntity): Promise<UserEntity> {
@@ -49,14 +55,27 @@ class UserRepository implements Repository {
 		});
 	}
 
-	public delete(): ReturnType<Repository["delete"]> {
-		return Promise.resolve(true);
+	public async delete(id: number): Promise<boolean> {
+		const rowsDeleted = await this.userModel.query().deleteById(id);
+
+		return Boolean(rowsDeleted);
+	}
+
+	public async deleteUserTaskDays(userId: number): Promise<boolean> {
+		const rowsDeleted = await this.userTaskDaysModel
+			.query()
+			.delete()
+			.where({ userId });
+
+		return Boolean(rowsDeleted);
 	}
 
 	public async find(id: number): Promise<null | UserEntity> {
 		const user = await this.userModel
 			.query()
-			.withGraphFetched(RelationName.USER_DETAILS)
+			.withGraphFetched(
+				`[${RelationName.USER_DETAILS}, ${RelationName.USER_TASK_DAYS}]`,
+			)
 			.findById(id);
 
 		return user
@@ -65,9 +84,13 @@ class UserRepository implements Repository {
 					email: user.email,
 					id: user.id,
 					name: user.userDetails.name,
+					notificationFrequency: user.notificationFrequency,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
+					userTaskDays: user.userTaskDays.map(
+						(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+					),
 				})
 			: null;
 	}
@@ -75,7 +98,9 @@ class UserRepository implements Repository {
 	public async findAll(): Promise<UserEntity[]> {
 		const users = await this.userModel
 			.query()
-			.withGraphFetched(RelationName.USER_DETAILS);
+			.withGraphFetched(
+				`[${RelationName.USER_DETAILS}, ${RelationName.USER_TASK_DAYS}]`,
+			);
 
 		return users.map((user) =>
 			UserEntity.initialize({
@@ -83,9 +108,13 @@ class UserRepository implements Repository {
 				email: user.email,
 				id: user.id,
 				name: user.userDetails.name,
+				notificationFrequency: user.notificationFrequency,
 				passwordHash: user.passwordHash,
 				passwordSalt: user.passwordSalt,
 				updatedAt: user.updatedAt,
+				userTaskDays: user.userTaskDays.map(
+					(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+				),
 			}),
 		);
 	}
@@ -93,9 +122,11 @@ class UserRepository implements Repository {
 	public async findByEmail(email: string): Promise<null | UserEntity> {
 		const user = await this.userModel
 			.query()
-			.withGraphFetched(RelationName.USER_DETAILS)
 			.where({ email })
-			.first();
+			.first()
+			.withGraphFetched(
+				`[${RelationName.USER_DETAILS}, ${RelationName.USER_TASK_DAYS}]`,
+			);
 
 		return user
 			? UserEntity.initialize({
@@ -103,9 +134,13 @@ class UserRepository implements Repository {
 					email: user.email,
 					id: user.id,
 					name: user.userDetails.name,
+					notificationFrequency: user.notificationFrequency,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
+					userTaskDays: user.userTaskDays.map(
+						(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+					),
 				})
 			: null;
 	}
@@ -120,7 +155,10 @@ class UserRepository implements Repository {
 		const updatedUserDetails = await userDetails
 			?.$query()
 			.patchAndFetch(payload);
-		const user = await this.userModel.query().findById(id);
+		const user = await this.userModel
+			.query()
+			.findById(id)
+			.withGraphFetched(RelationName.USER_TASK_DAYS);
 
 		return user && updatedUserDetails
 			? UserEntity.initialize({
@@ -128,9 +166,13 @@ class UserRepository implements Repository {
 					email: user.email,
 					id: user.id,
 					name: updatedUserDetails.name,
+					notificationFrequency: updatedUserDetails.notificationFrequency,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
+					userTaskDays: user.userTaskDays.map(
+						(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+					),
 				})
 			: null;
 	}
@@ -153,6 +195,19 @@ class UserRepository implements Repository {
 			passwordSalt: user.passwordSalt,
 			updatedAt: user.updatedAt,
 		});
+	}
+
+	public async updateUserTaskDays(
+		id: number,
+		userTaskDays: number[],
+	): Promise<void> {
+		await this.deleteUserTaskDays(id);
+
+		await this.userTaskDaysModel.query().insert(
+			userTaskDays.map((day) => {
+				return { dayOfWeek: day, userId: id };
+			}),
+		);
 	}
 }