+++ apps/frontend/src/pages/quiz/libs/types/quiz-form-values.types.ts
@@ -1,0 +1,5 @@
+type QuizFormValues = {
+	[key: string]: string;
+};
+
+export { type QuizFormValues };