+++ apps/mobile/src/libs/types/types.ts
@@ -5,8 +5,18 @@ export {
 	type NavigationScreenProperties,
 	type RootNavigationParameterList,
 } from "./navigation/navigation";
+export {
+	type RadioGroupItemStyle,
+	type RadioGroupOption,
+} from "./radio-group/types";
 export { type WheelDataItem } from "./wheel/wheel";
 export { type NativeStackNavigationProp } from "@react-navigation/native-stack";
+export {
+	type Control,
+	type FieldErrors,
+	type FieldPath,
+	type FieldValues,
+} from "react-hook-form";
 export {
 	type ImageSourcePropType,
 	type StyleProp,
@@ -14,7 +24,9 @@ export {
 	type ViewStyle,
 } from "react-native";
 export { type LinearGradientProps } from "react-native-linear-gradient";
+export { type RadioButtonProps } from "react-native-radio-buttons-group";
 export {
+	type OnboardingQuestionResponseDto,
 	type ServerErrorDetail,
 	type ServerErrorResponse,
 	type ValidationSchema,