+++ apps/mobile/src/screens/onboarding/libs/constants/onboarding-form-default-values.constant.ts
@@ -1,0 +1,7 @@
+import { type OnboardingFormValues } from "../types/types.js";
+
+const ONBOARDING_FORM_DEFAULT_VALUES: OnboardingFormValues = {
+	answer: "",
+};
+
+export { ONBOARDING_FORM_DEFAULT_VALUES };